+++ src/codeJeu/Monstre.java
@@ -2,7 +2,7 @@ package codeJeu;
 
 public class Monstre extends JoueurActif {
 	//constructeur monstre
-    public Monstre(double x, double y, int pdv, int force) {
+    public Monstre(int x, int y, int pdv, int force) {
 		super(x, y, pdv, force);
 	}
